abstraction of service framework and run type
