+++ codegen/smithy-kotlin-codegen/src/main/kotlin/software/amazon/smithy/kotlin/codegen/service/ServiceStubConfigurations.kt
@@ -1,5 +1,9 @@
 package software.amazon.smithy.kotlin.codegen.service
 
+import software.amazon.smithy.build.FileManifest
+import software.amazon.smithy.kotlin.codegen.core.GenerationContext
+import software.amazon.smithy.kotlin.codegen.core.KotlinDelegator
+
 enum class ServiceFramework(val value: String) {
     KTOR("ktor"),
     ;
@@ -10,6 +14,16 @@ enum class ServiceFramework(val value: String) {
         fun fromValue(value: String): ServiceFramework = when (value.lowercase()) {
             "ktor" -> KTOR
             else -> throw IllegalArgumentException("$value is not a valid ServerFramework value, expected $KTOR")
+        }
+    }
+
+    internal fun getServiceFrameworkGenerator(
+        ctx: GenerationContext,
+        delegator: KotlinDelegator,
+        fileManifest: FileManifest,
+    ): AbstractStubGenerator {
+        when (this) {
+            KTOR -> return KtorStubGenerator(ctx, delegator, fileManifest)
         }
     }
 }
